+++ source/extensions/lightspeed.trex.packaging.widget/lightspeed/trex/packaging/widget/tests/__init__.py
@@ -15,4 +15,4 @@
 * limitations under the License.
 """
 
-from .unit.test_visibility_filtering import *  # noqa: F401, F403
+from .unit.test_subscription_lifecycle import TestPackagingPaneSubscriptionLifecycle